Open files from the Files inspector as center editor tabs
Files now behaves like VS Code's explorer: a single click opens a preview
tab (italic title, one per session, re-targeted by the next click); a
double-click in the tree or on the tab, or editing the file, pins it. A
file that is already open is focused instead of duplicated.

- New `editor` panel type (shared/types/panels.ts), closable, never in the
  `+` menu, exempt from the closed-panel-type record.
- FileEditor's Monaco half extracted to FileEditorView; the Files tab is
  tree-only and highlights the active editor tab's file.
- openFileInEditor service owns preview/pin/focus and layout insertion;
  Review "Open in editor" and terminal file links use it (pinned).
- pane-drive: file:/dfile:/dtab:/tabs steps.
