+++ frontend/src/components/panels/editor/FileEditor.tsx
@@ -1,17 +1,8 @@
 import { useState, useEffect, useCallback, useRef, useMemo } from 'react';
-import Editor from '@monaco-editor/react';
-import type * as monaco from 'monaco-editor';
-import { ChevronRight, ChevronDown, File, Folder, RefreshCw, Plus, Trash2, FolderPlus, Search, X, Eye, Code, Copy, FolderOpen, Pencil, Clipboard, ClipboardPaste, CopyPlus } from 'lucide-react';
+import { ChevronRight, ChevronDown, File, Folder, RefreshCw, Plus, Trash2, FolderPlus, Search, X, Eye, Copy, FolderOpen, Pencil, Clipboard, ClipboardPaste, CopyPlus } from 'lucide-react';
 import { useTree } from '@headless-tree/react';
 import { asyncDataLoaderFeature, selectionFeature, hotkeysCoreFeature, expandAllFeature } from '@headless-tree/core';
 import type { ItemInstance } from '@headless-tree/core';
-import { MonacoErrorBoundary } from '../../MonacoErrorBoundary';
-import { isLightTheme, useTheme } from '../../../contexts/ThemeContext';
-import { debounce } from '../../../utils/debounce';
-import { devLog } from '../../../utils/console';
-import { MarkdownPreview } from '../../MarkdownPreview';
-import { NotebookPreview } from './NotebookPreview';
-import { useResizablePanel } from '../../../hooks/useResizablePanel';
 import { useCommittedRef } from '../../../hooks/useCommittedRef';
 import { ExplorerPanelState } from '../../../../../shared/types/panels';
 import { isMac, isWindows } from '../../../utils/platformUtils';
@@ -20,14 +11,10 @@ import { TerminalPopover, PopoverButton } from '../../terminal/TerminalPopover';
 import { areKeyboardShortcutsEnabled, useConfigStore } from '../../../stores/configStore';
 import { LiveRegion } from '../../ui/LiveRegion';
 import { boundary, decodeBoundary } from '../../../../../shared/validation/boundaryDecoder';
-import type { BrowserPanelState, ToolPanel } from '../../../../../shared/types/panels';
-import { panelApi } from '../../../services/panelApi';
-import { usePanelStore } from '../../../stores/panelStore';
 import { isHtmlFile } from './htmlFile';
-interface LanguageByExtension {
-  [extension: string]: string;
-}
-
+import { previewHtmlFileInBrowser } from './previewHtmlFile';
+import { editorPanelState, openFileInEditor } from '../../../services/openFileInEditor';
+import { usePanelStore } from '../../../stores/panelStore';
 interface FileItem {
   name: string;
   path: string;
@@ -38,14 +25,6 @@ interface FileItem {
 
 const ROOT_ID = '\0root';
 const MAX_FILE_SIZE = 15 * 1024 * 1024; // 15MB
-const IMAGE_EXTENSIONS = new Set(['png', 'jpg', 'jpeg', 'gif', 'webp', 'ico', 'bmp']);
-const PDF_EXTENSIONS = new Set(['pdf']);
-
-const filePathResponseSchema = boundary.object({
-  success: boundary.boolean,
-  url: boundary.optional(boundary.string),
-  error: boundary.optional(boundary.string),
-});
 
 function containsEventTarget(container: Node, target: EventTarget | null): boolean {
   return target instanceof Node && container.contains(target);
@@ -54,6 +33,8 @@ function containsEventTarget(container: Node, target: EventTarget | null): boole
 interface HeadlessFileTreeProps {
   sessionId: string;
   onFileSelect: (file: FileItem | null) => void;
+  /** Double-click: open the file pinned (VS Code semantics). */
+  onFileOpen?: (file: FileItem) => void;
   onFileCreateSelect?: (filePath: string) => void;
   selectedPath: string | null;
   initialExpandedDirs?: string[];
@@ -66,6 +47,7 @@ interface HeadlessFileTreeProps {
 function HeadlessFileTree({
   sessionId,
   onFileSelect,
+  onFileOpen,
   onFileCreateSelect,
   selectedPath,
   initialExpandedDirs,
@@ -906,6 +888,7 @@ function HeadlessFileTree({
                 disabled={file.isDirectory}
                 className="flex min-w-0 flex-1 items-center px-2 py-1 text-left disabled:cursor-default"
                 onClick={() => onFileSelect(file)}
+                onDoubleClick={() => onFileOpen?.(file)}
               >
                 {file.isDirectory ? (
                   <Folder className="w-4 h-4 mr-2 text-interactive flex-shrink-0" />
@@ -1032,7 +1015,10 @@ function HeadlessFileTree({
                   }
                 }
               }}
-              onDoubleClick={(e) => e.preventDefault()}
+              onDoubleClick={(e) => {
+                e.preventDefault();
+                if (!isFolder) onFileOpen?.(data);
+              }}
               onContextMenu={(e) => {
                 e.preventDefault();
                 e.stopPropagation();
@@ -1219,716 +1205,80 @@ function HeadlessFileTree({
 
 interface FileEditorProps {
   sessionId: string;
-  initialFilePath?: string;
   initialState?: ExplorerPanelState;
-  onFileChange?: (filePath: string | undefined, isDirty: boolean) => void;
   onStateChange?: (state: Partial<ExplorerPanelState>) => void;
 }
 
-export function FileEditor({ 
-  sessionId, 
-  initialFilePath,
-  initialState,
-  onFileChange,
-  onStateChange 
-}: FileEditorProps) {
-  devLog.debug('[FileEditor] Mounting with:', {
-    sessionId,
-    initialFilePath,
-    initialState,
-    hasOnStateChange: !!onStateChange
-  });
-  
-  const [selectedFile, setSelectedFile] = useState<FileItem | null>(null);
-  const selectedFilePathRef = useRef<string | null>(null);
-  useEffect(() => {
-    selectedFilePathRef.current = selectedFile?.path ?? null;
-  }, [selectedFile?.path]);
-  const mountedRef = useRef(true);
-  useEffect(() => {
-    mountedRef.current = true;
-    return () => {
-      mountedRef.current = false;
-    };
-  }, []);
-  const [fileContent, setFileContent] = useState<string>('');
-  const [originalContent, setOriginalContent] = useState<string>('');
-  const [loading, setLoading] = useState(false);
+/**
+ * The Files inspector: a file tree whose clicks open center editor tabs.
+ * Single-click previews, double-click pins (VS Code semantics); the row of
+ * the active editor tab's file is highlighted.
+ */
+export function FileEditor({ sessionId, initialState, onStateChange }: FileEditorProps) {
   const [error, setError] = useState<string | null>(null);
-  const [viewMode, setViewMode] = useState<'edit' | 'preview'>('edit');
-  const [gitStatus, setGitStatus] = useState<'clean' | 'modified' | 'untracked'>('clean');
-  const [binaryBlobUrl, setBinaryBlobUrl] = useState<string | null>(null);
-  const binaryBlobUrlRef = useRef<string | null>(null);
-  const editorRef = useRef<monaco.editor.IStandaloneCodeEditor | null>(null);
-  const monacoRef = useRef<typeof monaco | null>(null);
-  const pendingEditorFocusPathRef = useRef<string | null>(null);
-  const addPanel = usePanelStore((state) => state.addPanel);
-  const setActivePanel = usePanelStore((state) => state.setActivePanel);
-  const updatePanelState = usePanelStore((state) => state.updatePanelState);
+  const pendingFocusPathRef = useRef<string | null>(null);
 
-  // Keep ref in sync and clean up blob URLs to prevent memory leaks
-  useEffect(() => {
-    binaryBlobUrlRef.current = binaryBlobUrl;
-    return () => {
-      if (binaryBlobUrl) URL.revokeObjectURL(binaryBlobUrl);
-    };
-  }, [binaryBlobUrl]);
-
-  const { theme } = useTheme();
-  const isDarkMode = !isLightTheme(theme);
-  const hasUnsavedChanges = fileContent !== originalContent;
-  
-  // Wrap onResize callback to avoid recreating
-  const handleTreeResize = useCallback((width: number) => {
-    devLog.debug('[FileEditor] Tree resized to:', width);
-    if (onStateChange) {
-      onStateChange({ fileTreeWidth: width });
-    }
-  }, [onStateChange]);
-  
-  // Add resizable hook for file tree column
-  const { width: fileTreeWidth, startResize } = useResizablePanel({
-    defaultWidth: initialState?.fileTreeWidth || 256,  // Use saved width or default
-    minWidth: 200,
-    maxWidth: 400,
-    storageKey: 'pane-file-tree-width',
-    onResize: handleTreeResize
+  const activeEditorPath = usePanelStore((state) => {
+    const activeId = state.activePanels[sessionId];
+    const active = (state.panels[sessionId] || []).find((panel) => panel.id === activeId);
+    return active ? editorPanelState(active)?.filePath ?? null : null;
   });
-  
-  // Check if this is a markdown file
-  const isMarkdownFile = useMemo(() => {
-    if (!selectedFile) return false;
-    const ext = selectedFile.path.split('.').pop()?.toLowerCase();
-    return ext === 'md' || ext === 'markdown';
-  }, [selectedFile]);
 
-  // Check if this is a notebook file
-  const isNotebookFile = useMemo(() => {
-    if (!selectedFile) return false;
-    const ext = selectedFile.path.split('.').pop()?.toLowerCase();
-    return ext === 'ipynb';
-  }, [selectedFile]);
+  const openFile = useCallback(async (file: FileItem | null, pin: boolean) => {
+    if (!file || file.isDirectory) return;
+    setError(null);
+    try {
+      await openFileInEditor({ sessionId, filePath: file.path, pin });
+      if (pendingFocusPathRef.current === file.path) {
+        pendingFocusPathRef.current = null;
+        window.dispatchEvent(new CustomEvent('editor-panel:reveal', {
+          detail: { filePath: file.path, cursorPosition: { line: 1, column: 1 } },
+        }));
+      }
+    } catch (openError) {
+      setError(openError instanceof Error ? openError.message : 'Failed to open file');
+    }
+  }, [sessionId]);
 
-  const isImageFile = useMemo(() => {
-    if (!selectedFile) return false;
-    const ext = selectedFile.path.split('.').pop()?.toLowerCase() || '';
-    return IMAGE_EXTENSIONS.has(ext);
-  }, [selectedFile]);
-
-  const isPdfFile = useMemo(() => {
-    if (!selectedFile) return false;
-    const ext = selectedFile.path.split('.').pop()?.toLowerCase() || '';
-    return PDF_EXTENSIONS.has(ext);
-  }, [selectedFile]);
-
-  const isBinaryPreview = isImageFile || isPdfFile;
+  const handleFileSelect = useCallback((file: FileItem | null) => { void openFile(file, false); }, [openFile]);
+  const handleFileOpen = useCallback((file: FileItem) => { void openFile(file, true); }, [openFile]);
 
   const previewHtmlFile = useCallback(async (filePath: string) => {
     setError(null);
     try {
-      const result = decodeBoundary(
-        await window.electronAPI.invoke('file:getPath', { sessionId, filePath }),
-        filePathResponseSchema,
-      );
-      if (!result.success || !result.url) {
-        throw new Error(result.error || 'Failed to resolve HTML preview URL');
-      }
-
-      const panels = usePanelStore.getState().getSessionPanels(sessionId);
-      const existingPanel = panels.find((candidate) => candidate.type === 'browser');
-      let browserPanel: ToolPanel;
-
-      if (existingPanel) {
-        // SAFETY: The browser panel type discriminator establishes BrowserPanelState.
-        const existingCustomState = (existingPanel.state.customState ?? {}) as BrowserPanelState;
-        browserPanel = {
-          ...existingPanel,
-          title: filePath.split('/').pop() || 'Browser',
-          state: {
-            ...existingPanel.state,
-            customState: { ...existingCustomState, currentUrl: result.url },
-          },
-        };
-        await panelApi.updatePanel(browserPanel.id, {
-          title: browserPanel.title,
-          state: browserPanel.state,
-        });
-        updatePanelState(browserPanel);
-      } else {
-        browserPanel = await panelApi.createPanel({
-          sessionId,
-          type: 'browser',
-          title: filePath.split('/').pop() || 'Browser',
-          initialState: { customState: { currentUrl: result.url } },
-        });
-        addPanel(browserPanel);
-      }
-
-      setActivePanel(sessionId, browserPanel.id);
-      await panelApi.setActivePanel(sessionId, browserPanel.id);
-      window.dispatchEvent(new CustomEvent('browser-panel:navigate', {
-        detail: { url: result.url, sessionId },
-      }));
+      await previewHtmlFileInBrowser(sessionId, filePath);
     } catch (previewError) {
       setError(previewError instanceof Error ? previewError.message : 'Failed to preview HTML file');
     }
-  }, [sessionId, addPanel, setActivePanel, updatePanelState]);
+  }, [sessionId]);
 
-  const loadFile = useCallback(async (file: FileItem | null) => {
-    if (!file || file.isDirectory) return;
-
-    setLoading(true);
-    setError(null);
-    setGitStatus('clean');
-    try {
-      // Binary file detection — render as image/PDF preview instead of Monaco
-      const ext = file.path.split('.').pop()?.toLowerCase() || '';
-      const isImage = IMAGE_EXTENSIONS.has(ext);
-      const isPdf = PDF_EXTENSIONS.has(ext);
-
-      if (isImage || isPdf) {
-        const result = await window.electronAPI.invoke('file:read-binary', {
-          sessionId,
-          filePath: file.path,
-        });
-        if (result.success && result.contentBase64) {
-          // Revoke previous blob URL via ref (avoids stale closure from useCallback)
-          if (binaryBlobUrlRef.current) URL.revokeObjectURL(binaryBlobUrlRef.current);
-
-          const mimeType = isImage
-            ? `image/${ext === 'jpg' ? 'jpeg' : ext === 'ico' ? 'x-icon' : ext}`
-            : 'application/pdf';
-          const byteChars = atob(result.contentBase64);
-          const byteArray = new Uint8Array(byteChars.length);
-          for (let i = 0; i < byteChars.length; i++) {
-            byteArray[i] = byteChars.charCodeAt(i);
-          }
-          const blob = new Blob([byteArray], { type: mimeType });
-          setBinaryBlobUrl(URL.createObjectURL(blob));
-          setFileContent('');
-          setOriginalContent('');
-        } else {
-          // Binary read failed — show error instead of blank/stale preview
-          setBinaryBlobUrl(null);
-          setError(result.error || 'Failed to load binary file');
-        }
-        setSelectedFile(file);
-        setViewMode('edit');
-        setLoading(false);
-        onFileChange?.(file.path, false);
-        onStateChange?.({ filePath: file.path });
-
-        // Check git status for binary files too
-        window.electronAPI.invoke('git:file-status', sessionId, file.path).then((statusResult: { success: boolean; data?: { status: 'clean' | 'modified' | 'untracked' } }) => {
-          if (statusResult.success && statusResult.data) {
-            setGitStatus(statusResult.data.status);
-          }
-        });
-        return;
-      }
-
-      const result = await window.electronAPI.invoke('file:read', {
-        sessionId,
-        filePath: file.path
-      });
-
-      if (result.success) {
-        setBinaryBlobUrl(null);
-        setFileContent(result.content);
-        setOriginalContent(result.content);
-        setSelectedFile(file);
-        setViewMode('edit'); // Reset to edit mode when opening a new file
-        if (pendingEditorFocusPathRef.current === file.path) {
-          window.setTimeout(() => editorRef.current?.focus(), 100);
-        }
-        
-        // Notify parent about file change
-        if (onFileChange) {
-          onFileChange(file.path, false);
-        }
-        
-        // After loading new file, we need to restore its position
-        // This happens in handleEditorMount when editor re-renders
-        // But we also need to tell parent the file path changed
-        if (onStateChange) {
-          onStateChange({ 
-            filePath: file.path,
-            isDirty: false 
-          });
-        }
-        
-        // If we have saved position for this file, restore it
-        // The actual restoration happens in handleEditorMount
-        // but we need to trigger a re-render with the right state
-        if (editorRef.current && initialState?.filePath === file.path) {
-          const monacoEditor = editorRef.current;
-          
-          // Restore cursor position
-          if (initialState.cursorPosition && monacoEditor.setPosition) {
-            const { line, column } = initialState.cursorPosition;
-            setTimeout(() => {
-              monacoEditor.setPosition({
-                lineNumber: line,
-                column: column
-              });
-              monacoEditor.revealPositionInCenter({
-                lineNumber: line,
-                column: column
-              });
-            }, 50);
-          }
-          
-          // Restore scroll position
-          if (initialState.scrollPosition !== undefined && monacoEditor.setScrollTop) {
-            const scrollPos = initialState.scrollPosition;
-            setTimeout(() => {
-              monacoEditor.setScrollTop(scrollPos);
-            }, 100);
-          }
-        }
-
-        // Check git status for this file
-        window.electronAPI.invoke('git:file-status', sessionId, file.path).then((statusResult: { success: boolean; data?: { status: 'clean' | 'modified' | 'untracked' } }) => {
-          if (statusResult.success && statusResult.data) {
-            setGitStatus(statusResult.data.status);
-          }
-        });
-      } else {
-        setError(result.error);
-      }
-    } catch (err) {
-      setError(err instanceof Error ? err.message : 'Failed to load file');
-    } finally {
-      setLoading(false);
-    }
-  }, [sessionId, onFileChange, onStateChange, initialState, binaryBlobUrlRef]);
-
-  const selectedFilePath = selectedFile?.path;
-
-  useEffect(() => {
-    if (!selectedFilePath || pendingEditorFocusPathRef.current !== selectedFilePath) return;
-    const focusTimer = window.setTimeout(() => {
-      editorRef.current?.focus();
-      pendingEditorFocusPathRef.current = null;
-    }, 100);
-    return () => window.clearTimeout(focusTimer);
-  }, [selectedFilePath]);
-
-
-  const handleEditorMount = (editor: monaco.editor.IStandaloneCodeEditor, monacoInstance: typeof monaco) => {
-    editorRef.current = editor;
-    monacoRef.current = monacoInstance;
-    
-    // Now we have properly typed Monaco editor
-    const monacoEditor = editor;
-    
-    // Track cursor position changes with debouncing
-    const saveCursorPosition = debounce((position: { lineNumber: number; column: number }) => {
-      if (onStateChange) {
-        onStateChange({
-          cursorPosition: {
-            line: position.lineNumber,
-            column: position.column
-          }
-        });
-      }
-    }, 500); // Debounce cursor position saves
-    
-    // Track scroll position changes with debouncing
-    const saveScrollPosition = debounce((scrollTop: number) => {
-      if (onStateChange) {
-        onStateChange({
-          scrollPosition: scrollTop
-        });
-      }
-    }, 500); // Debounce scroll position saves
-    
-    // Listen for cursor position changes
-    monacoEditor.onDidChangeCursorPosition?.((e: monaco.editor.ICursorPositionChangedEvent) => {
-      saveCursorPosition(e.position);
-    });
-    
-    // Listen for scroll position changes
-    monacoEditor.onDidScrollChange?.((e: { scrollTop?: number; scrollLeft?: number }) => {
-      if (e.scrollTop !== undefined) {
-        saveScrollPosition(e.scrollTop);
-      }
-    });
-    
-    // Restore cursor and scroll position if available
-    if (initialState?.cursorPosition && monacoEditor.setPosition) {
-      const { line, column } = initialState.cursorPosition;
-      setTimeout(() => {
-        monacoEditor.setPosition({
-          lineNumber: line,
-          column: column
-        });
-        monacoEditor.revealPositionInCenter({
-          lineNumber: line,
-          column: column
-        });
-      }, 50); // Small delay to ensure editor is ready
-    }
-    
-    if (initialState?.scrollPosition !== undefined && monacoEditor.setScrollTop) {
-      // Delay to ensure editor is fully rendered and content is loaded
-      const scrollPos = initialState.scrollPosition;
-      setTimeout(() => {
-        monacoEditor.setScrollTop(scrollPos);
-      }, 100);
-    }
-  };
-
-  // Auto-save functionality
-  const pendingAutoSaveFilePathRef = useRef<string | null>(null);
-  const autoSave = useMemo(
-    () => debounce(async (file: FileItem, content: string) => {
-      pendingAutoSaveFilePathRef.current = null;
-      try {
-        const result = await window.electronAPI.invoke('file:write', {
-          sessionId,
-          filePath: file.path,
-          content,
-        });
-        
-        if (result.success) {
-          const isActiveFile = mountedRef.current && selectedFilePathRef.current === file.path;
-          if (isActiveFile) {
-            setOriginalContent(content);
-            onFileChange?.(file.path, false);
-            onStateChange?.({
-              filePath: file.path,
-              isDirty: false,
-            });
-          }
-
-          // Re-check git status after save
-          window.electronAPI.invoke('git:file-status', sessionId, file.path).then((statusResult: { success: boolean; data?: { status: 'clean' | 'modified' | 'untracked' } }) => {
-            if (mountedRef.current && selectedFilePathRef.current === file.path && statusResult.success && statusResult.data) {
-              setGitStatus(statusResult.data.status);
-            }
-          });
-        } else if (mountedRef.current) {
-          setError(result.error);
-        }
-      } catch (err) {
-        if (mountedRef.current) {
-          setError(err instanceof Error ? err.message : 'Failed to auto-save file');
-        }
-      }
-    }, 1000), // Auto-save after 1 second of inactivity
-    [sessionId, onFileChange, onStateChange]
-  );
-
-  useEffect(() => () => autoSave.flush(), [autoSave]);
-
-  const handleEditorChange = (value: string | undefined) => {
-    const content = value || '';
-    setFileContent(content);
-
-    if (selectedFile && !selectedFile.isDirectory) {
-      const isDirty = content !== originalContent;
-      onFileChange?.(selectedFile.path, isDirty);
-      if (isDirty) {
-        if (pendingAutoSaveFilePathRef.current && pendingAutoSaveFilePathRef.current !== selectedFile.path) {
-          autoSave.flush();
-        }
-        pendingAutoSaveFilePathRef.current = selectedFile.path;
-        autoSave(selectedFile, content);
-      } else if (pendingAutoSaveFilePathRef.current === selectedFile.path) {
-        autoSave.cancel();
-        pendingAutoSaveFilePathRef.current = null;
-      }
-    }
-  };
-
-  // Re-check git status when git operations complete (e.g. commit from diff panel or terminal)
-  useEffect(() => {
-    if (!selectedFile) return;
-    const handlePanelEvent = (event: Event) => {
-      if (!(event instanceof CustomEvent)) return;
-      const { type } = event.detail || {};
-      if (type === 'git:operation_completed' || type === 'diff:refreshed' || type === 'terminal:command_executed' || type === 'files:changed') {
-        window.electronAPI.invoke('git:file-status', sessionId, selectedFile.path).then((statusResult: { success: boolean; data?: { status: 'clean' | 'modified' | 'untracked' } }) => {
-          if (statusResult.success && statusResult.data) {
-            setGitStatus(statusResult.data.status);
-          }
-        });
-      }
-    };
-    window.addEventListener('panel:event', handlePanelEvent);
-    return () => window.removeEventListener('panel:event', handlePanelEvent);
-  }, [selectedFile, sessionId]);
-  
-  // Load initial file if provided
-  useEffect(() => {
-    if (initialFilePath && !selectedFile) {
-      const file: FileItem = {
-        name: initialFilePath.split('/').pop() || '',
-        path: initialFilePath,
-        isDirectory: false
-      };
-      loadFile(file);
-    }
-  }, [initialFilePath, selectedFile, loadFile]);
-
-  // Memoize the tree state change handler to prevent infinite loops
   const handleTreeStateChange = useCallback((treeState: { expandedDirs: string[]; searchQuery: string; showSearch: boolean }) => {
-    devLog.debug('[FileEditor] handleTreeStateChange called with:', treeState);
-    if (onStateChange) {
-      devLog.debug('[FileEditor] Calling onStateChange');
-      onStateChange({
-        expandedDirs: treeState.expandedDirs,
-        searchQuery: treeState.searchQuery,
-        showSearch: treeState.showSearch
-      });
-    } else {
-      devLog.debug('[FileEditor] No onStateChange callback');
-    }
+    onStateChange?.(treeState);
   }, [onStateChange]);
-  
-  // Cleanup effect for Monaco editor models
-  useEffect(() => {
-    return () => {
-      // Cleanup Monaco editor models when component unmounts or file changes
-      try {
-        const model = editorRef.current?.getModel();
-        if (model) {
-          devLog.debug('[FileEditor] Disposing Monaco model');
-          model.dispose();
-        }
-      } catch (error) {
-        console.warn('[FileEditor] Error during Monaco cleanup:', error);
-      }
-    };
-  }, [selectedFile?.path]); // Run cleanup when file changes
 
   return (
-    <div className="pane-explorer-split h-full w-full min-w-0 flex overflow-hidden" data-empty={selectedFile ? undefined : 'true'}>
-      <div 
-        className="pane-explorer-tree bg-surface-secondary border-r border-border-primary relative flex-shrink-0 max-w-[45%]"
-        style={{ width: `${fileTreeWidth}px` }}
-      >
+    <div className="pane-explorer-tree h-full w-full min-w-0 flex flex-col overflow-hidden bg-surface-secondary">
+      {error && (
+        <div role="alert" className="px-3 py-1.5 bg-status-error/20 text-status-error text-xs">
+          {error}
+        </div>
+      )}
+      <div className="flex-1 min-h-0">
         <HeadlessFileTree
           sessionId={sessionId}
-          onFileSelect={loadFile}
+          onFileSelect={handleFileSelect}
+          onFileOpen={handleFileOpen}
           onFileCreateSelect={(filePath) => {
-            pendingEditorFocusPathRef.current = filePath;
+            pendingFocusPathRef.current = filePath;
           }}
-          selectedPath={selectedFile?.path || null}
+          selectedPath={activeEditorPath}
           initialExpandedDirs={initialState?.expandedDirs}
           initialSearchQuery={initialState?.searchQuery}
           initialShowSearch={initialState?.showSearch}
           onTreeStateChange={handleTreeStateChange}
           onHtmlPreview={previewHtmlFile}
         />
-        
-        {/* Resize handle */}
-        <div
-          className="absolute top-0 right-0 w-1 h-full cursor-col-resize group z-10"
-          onMouseDown={startResize}
-        >
-          {/* Larger grab area */}
-          <div className="absolute -left-2 -right-2 top-0 bottom-0" />
-        </div>
-      </div>
-      <div className="pane-explorer-editor flex-1 min-w-0 flex flex-col overflow-hidden">
-        {selectedFile ? (
-          <>
-            <div className="flex items-center justify-between px-4 py-2 bg-surface-secondary border-b border-border-primary">
-              <div className="flex min-w-0 items-center gap-2">
-                {isHtmlFile(selectedFile.path) && (
-                  <button
-                    type="button"
-                    onClick={() => previewHtmlFile(selectedFile.path)}
-                    className="flex flex-shrink-0 items-center gap-1 rounded-lg border border-border-primary bg-surface-tertiary px-2 py-1 text-xs font-medium text-text-secondary transition-colors hover:bg-surface-hover hover:text-text-primary"
-                    title="Preview file as HTML"
-                    aria-label="Preview file as HTML"
-                  >
-                    <Eye className="w-3 h-3" />
-                    Preview as HTML
-                  </button>
-                )}
-                <File className="w-4 h-4 text-text-tertiary" />
-                <span className="min-w-0 truncate text-sm text-text-primary">
-                  {selectedFile.path}
-                  {hasUnsavedChanges && <span className="text-status-warning ml-2">●</span>}
-                </span>
-                {gitStatus !== 'clean' && (
-                  <span className={`text-xs font-medium px-1.5 py-0.5 rounded ${
-                    gitStatus === 'untracked'
-                      ? 'bg-status-success text-text-on-status-success'
-                      : 'bg-interactive text-text-on-interactive'
-                  }`}>
-                    {gitStatus === 'untracked' ? 'U' : 'M'}
-                  </span>
-                )}
-              </div>
-              <div className="flex flex-shrink-0 items-center gap-2">
-                {/* Preview Toggle for Markdown/Notebook Files */}
-                {!isBinaryPreview && (isMarkdownFile || isNotebookFile) && (
-                  <div className="flex items-center rounded-lg border border-border-primary bg-surface-tertiary">
-                    <button
-                      onClick={() => setViewMode('edit')}
-                      className={`px-2 py-1 text-xs font-medium rounded-l-lg transition-colors flex items-center gap-1 ${
-                        viewMode === 'edit'
-                          ? 'bg-interactive text-text-on-interactive'
-                          : 'text-text-secondary hover:bg-surface-hover'
-                      }`}
-                      title="Edit mode"
-                    >
-                      <Code className="w-3 h-3" />
-                      Edit
-                    </button>
-                    <button
-                      onClick={() => setViewMode('preview')}
-                      className={`px-2 py-1 text-xs font-medium rounded-r-lg transition-colors flex items-center gap-1 ${
-                        viewMode === 'preview'
-                          ? 'bg-interactive text-text-on-interactive'
-                          : 'text-text-secondary hover:bg-surface-hover'
-                      }`}
-                      title="Preview mode"
-                    >
-                      <Eye className="w-3 h-3" />
-                      Preview
-                    </button>
-                  </div>
-                )}
-                {!isBinaryPreview && (
-                  <div className="flex items-center gap-2 text-sm">
-                    {hasUnsavedChanges ? (
-                      <>
-                        <div className="w-2 h-2 bg-status-warning rounded-full animate-pulse" />
-                        <span className="text-status-warning">Auto-saving...</span>
-                      </>
-                    ) : (
-                      <>
-                        <div className="w-2 h-2 bg-status-success rounded-full" />
-                        <span className="text-status-success">All changes saved</span>
-                      </>
-                    )}
-                  </div>
-                )}
-              </div>
-            </div>
-            {error && (
-              <div role="alert" className="px-4 py-2 bg-status-error/20 text-status-error text-sm">
-                Error: {error}
-              </div>
-            )}
-            <div className="flex-1 min-w-0 overflow-hidden">
-              {viewMode === 'preview' && isMarkdownFile ? (
-                <div className="h-full overflow-auto bg-bg-primary">
-                  <MarkdownPreview
-                    content={fileContent}
-                    className="min-h-full"
-                    id={`file-editor-preview-${sessionId}-${selectedFile.path.replace(/[^a-zA-Z0-9]/g, '-')}`}
-                  />
-                </div>
-              ) : viewMode === 'preview' && isNotebookFile ? (
-                <div className="h-full overflow-auto bg-bg-primary">
-                  <NotebookPreview
-                    content={fileContent}
-                    className="min-h-full"
-                  />
-                </div>
-              ) : isBinaryPreview && !binaryBlobUrl && !error ? (
-                <div className="flex items-center justify-center h-full bg-surface-primary">
-                  <div className="animate-pulse flex flex-col items-center gap-3">
-                    <div className="w-48 h-48 bg-surface-tertiary rounded" />
-                    <div className="w-32 h-3 bg-surface-tertiary rounded" />
-                  </div>
-                </div>
-              ) : isImageFile && binaryBlobUrl ? (
-                <div className="flex items-center justify-center h-full bg-surface-primary p-4 overflow-auto">
-                  <img
-                    src={binaryBlobUrl}
-                    alt={selectedFile?.path.split('/').pop() || 'Image'}
-                    className="max-w-full max-h-full object-contain rounded"
-                  />
-                </div>
-              ) : isPdfFile && binaryBlobUrl ? (
-                <object
-                  data={binaryBlobUrl}
-                  type="application/pdf"
-                  className="w-full h-full"
-                >
-                  <div className="flex items-center justify-center h-full text-text-secondary">
-                    PDF preview not available.
-                  </div>
-                </object>
-              ) : (
-                <MonacoErrorBoundary>
-                  <Editor
-                    theme={isDarkMode ? 'vs-dark' : 'light'}
-                    value={fileContent}
-                    onChange={handleEditorChange}
-                    onMount={handleEditorMount}
-                    options={{
-                      minimap: { enabled: true },
-                      fontSize: 14,
-                      wordWrap: 'on',
-                      automaticLayout: true,
-                    }}
-                    language={getLanguageFromPath(selectedFile.path)}
-                  />
-                </MonacoErrorBoundary>
-              )}
-            </div>
-          </>
-        ) : (
-          <div className="flex-1 flex items-center justify-center text-text-secondary">
-            {loading ? 'Loading...' : 'Select a file to edit'}
-          </div>
-        )}
       </div>
     </div>
   );
-}
-
-function getLanguageFromPath(filePath: string): string {
-  const ext = filePath.split('.').pop()?.toLowerCase();
-  const languageMap: LanguageByExtension = {
-    js: 'javascript',
-    jsx: 'javascript',
-    ts: 'typescript',
-    tsx: 'typescript',
-    json: 'json',
-    ipynb: 'json',
-    md: 'markdown',
-    py: 'python',
-    rb: 'ruby',
-    go: 'go',
-    rs: 'rust',
-    cpp: 'cpp',
-    c: 'c',
-    h: 'c',
-    hpp: 'cpp',
-    java: 'java',
-    cs: 'csharp',
-    php: 'php',
-    html: 'html',
-    css: 'css',
-    scss: 'scss',
-    sass: 'sass',
-    less: 'less',
-    xml: 'xml',
-    yaml: 'yaml',
-    yml: 'yaml',
-    toml: 'toml',
-    ini: 'ini',
-    sh: 'shell',
-    bash: 'shell',
-    zsh: 'shell',
-    fish: 'shell',
-    ps1: 'powershell',
-    dockerfile: 'dockerfile',
-    makefile: 'makefile',
-    sql: 'sql',
-    graphql: 'graphql',
-    vue: 'vue',
-    svelte: 'svelte',
-  };
-  
-  return languageMap[ext || ''] || 'plaintext';
 }
