+++ scripts/pane-drive.mjs
@@ -41,6 +41,10 @@ try {
     if (s === 'dump') { console.log('URL', page.url()); console.log('TABS', await page.evaluate(() => Array.from(document.querySelectorAll('[role=tab]')).map(t => (t.getAttribute('aria-label') || t.textContent || '').trim()).join(' | '))); console.log('INSPECTOR', await page.evaluate(() => !!document.querySelector('[role=tablist][aria-label=Inspector]'))); console.log('SRC', await page.evaluate(async () => (await (await fetch('/src/components/SessionView.tsx')).text()).includes('isInspectorPanel'))); }
     if (s === 'dump2') { console.log('RES', await page.evaluate(() => performance.getEntriesByType('resource').map(r => r.name).filter(n => /SessionView|DetailPanel|InspectorTabs|main\.tsx|index-|assets\//.test(n)).slice(0,12).join('\n'))); console.log('SCRIPTS', await page.evaluate(() => Array.from(document.scripts).map(s => s.src).join(' '))); console.log('DETAIL', await page.evaluate(() => !!document.querySelector('.pane-detail-panel-vertical'))); }
     if (s === 'dump3') { await page.waitForTimeout(3000); console.log('LS', await page.evaluate(() => JSON.stringify(Object.fromEntries(Object.keys(localStorage).filter(k => /detail|inspector|layout/.test(k)).map(k => [k, localStorage.getItem(k)]))))); console.log('TABS3', await page.evaluate(() => Array.from(document.querySelectorAll('[role=tab]')).map(t => t.outerHTML.slice(0, 160)).join('\n'))); console.log('DW', await page.evaluate(() => document.querySelector(".pane-detail-panel-vertical")?.style.width)); }
+    if (s.startsWith('file:')) { const n = s.slice(5); await page.locator('.pane-explorer-tree').getByText(n, { exact: true }).first().click(); await page.waitForTimeout(2000); await shot(page, `e12-file-${n.replace(/[^a-z0-9]/gi,'_')}.png`); }
+    if (s.startsWith('dfile:')) { const n = s.slice(6); await page.locator('.pane-explorer-tree').getByText(n, { exact: true }).first().dblclick(); await page.waitForTimeout(2000); await shot(page, `e13-dfile-${n.replace(/[^a-z0-9]/gi,'_')}.png`); }
+    if (s.startsWith('dtab:')) { const n = s.slice(5); await page.getByRole('tab', { name: n, exact: true }).first().dblclick(); await page.waitForTimeout(1500); await shot(page, `e14-dtab-${n.replace(/[^a-z0-9]/gi,'_')}.png`); }
+    if (s === 'tabs') { console.log('TABS', await page.evaluate(() => Array.from(document.querySelectorAll('[role=tab]')).map(t => { const span = t.parentElement?.querySelector('span.truncate'); const italic = span ? getComputedStyle(span).fontStyle : '?'; return `${(t.getAttribute('aria-label') || '').trim()}[${italic}${t.getAttribute('aria-selected')==='true'?',active':''}]`; }).join(' | '))); }
     if (s === 'sidemenu') { await page.getByRole('button', { name: 'Sidebar menu' }).click(); await page.waitForTimeout(500); await shot(page, 'e11-sidemenu.png'); await page.keyboard.press('Escape'); }
     if (s === 'focus') { await page.locator('.xterm').first().click({ position: { x: 300, y: 200 } }); await page.waitForTimeout(300); }
     let n = 0;
